requirements.txt
django==1.7
djangorestframework==3.0
django-rest-swagger
django-filter
requests
gunicorn
psycopg2
mailchimp
pillow
django-autoslug

django-jenkins
django-extensions
sphinx
